Configure ackee-tracker for a production deploy test
We're exploring non-google analytics solutions.

Currently testing Ackee. The server is already setup,
now we need to start sending events from a client-side
tracker.

With this PR we're creating an instance of the Ackee
tracker and sending events.

Locally keeps failing with CORS issues despite the CORS
allowing for localhost. Deploying to prod to see what happens